requirements.txt
# # im cmd terminal nach dem Starten eingeben:
#C:\\ProgramData\\anaconda3\\Scripts\\activate.bat C:\\ProgramData\\anaconda3
#anaconda: conda activate myshinyenv_20240513
# pip install -r requirements.txt
openai 
pandas 
numpy 
python-dotenv